backend/requirements.txt
fastapi
uvicorn
feedparser
python-dotenv
requests